Re-enable limiting to Citation objects.
The original Altmetrics code limits the use of this module to Citation objects. Since other kinds of objects can contain DOIs and be cited/referenced with Altmetrics, I want to remove this limitation - but doing so generates warnings on Collections and other objects - trying to get property of non-object etc. The limitation is re-enabled until I can figure out a way around the warnings.